feat: remove cursor bundling from asdf-overlay-window
