+++ crates/dll/Cargo.toml
@@ -33,6 +33,9 @@ features = ["Win32_Security", "Win32_System_Diagnostics_Debug"]
 version = "0.1"
 features = ["max_level_trace", "release_max_level_off"]
 
+[build-dependencies]
+winres = "0.1.12"
+
 [target.'cfg(target_env = "msvc")'.build-dependencies]
 static_vcruntime = "3"
