+++ crates/dll/resources/cursors.rc
