+++ crates/dll/resources/cursors/aliasb.cur
