+++ crates/dll/resources/cursors/cell.cur
