+++ crates/dll/resources/cursors/col_resize.cur
