+++ crates/dll/resources/cursors/copy.cur
