+++ crates/dll/resources/cursors/hand_grab.cur
