+++ crates/dll/resources/cursors/hand_grabbing.cur
