+++ crates/dll/resources/cursors/pan_east.cur
