+++ crates/dll/resources/cursors/pan_middle.cur
