+++ crates/dll/resources/cursors/pan_middle_horizontal.cur
