+++ crates/dll/resources/cursors/pan_middle_vertical.cur
