+++ crates/dll/resources/cursors/pan_north.cur
