+++ crates/dll/resources/cursors/pan_north_east.cur
