+++ crates/dll/resources/cursors/pan_north_west.cur
