+++ crates/dll/resources/cursors/pan_south.cur
