+++ crates/dll/resources/cursors/pan_south_east.cur
