+++ crates/dll/resources/cursors/pan_south_west.cur
