+++ crates/dll/resources/cursors/pan_west.cur
