+++ crates/dll/resources/cursors/row_resize.cur
