+++ crates/dll/resources/cursors/vertical_text.cur
