+++ crates/dll/resources/cursors/zoom_in.cur
