+++ crates/dll/resources/cursors/zoom_out.cur
